refactor(test): split oracle tests and cleanup
