+++ tests/unit/test_process_lock.py
@@ -1,5 +1,8 @@
 """Process lock unit tests: contention, PID metadata, stable files."""
 
+import shutil
+import subprocess
+import sys
 from pathlib import Path
 
 import pytest
@@ -66,3 +69,106 @@ def test_import_does_not_lock():
     import services.process_lock as pl
 
     assert pl.AdvisoryFileLock is not None
+
+
+# ---------------------------------------------------------------------------
+# Process lock cross-process + Go interop
+# ---------------------------------------------------------------------------
+
+
+class TestProcessLockCrossProcess:
+    def test_cross_process_python_contention(self, tmp_path: Path):
+        lock_path = tmp_path / "config" / "locks" / "runtime.lock"
+        lock_path.parent.mkdir(parents=True)
+        holder = AdvisoryFileLock(lock_path)
+        holder.acquire()
+        try:
+            code = (
+                "from services.process_lock import AdvisoryFileLock, LockBusyError\n"
+                f"p = r'''{lock_path}'''\n"
+                "try:\n"
+                "    AdvisoryFileLock(p).acquire(timeout_seconds=None)\n"
+                "    print('ACQUIRED')\n"
+                "except LockBusyError:\n"
+                "    print('BUSY')\n"
+            )
+            proc = subprocess.run(
+                [sys.executable, "-c", code],
+                capture_output=True,
+                text=True,
+                cwd=str(Path(__file__).resolve().parents[2]),
+                timeout=15,
+            )
+            assert "BUSY" in proc.stdout
+        finally:
+            holder.release()
+
+    def test_python_go_lockhelper_interop(self, tmp_path: Path):
+        """Build lockhelper from source into temp binary; test both directions."""
+        root = Path(__file__).resolve().parents[2]
+        go = shutil.which("go")
+        if not go:
+            pytest.skip("Go toolchain unavailable; cannot build lockhelper")
+
+        helper_dir = tmp_path / "lockhelper_build"
+        helper_dir.mkdir()
+        helper = helper_dir / (
+            "lockhelper.exe" if sys.platform == "win32" else "lockhelper"
+        )
+        build = subprocess.run(
+            [go, "build", "-o", str(helper), "./cmd/lockhelper"],
+            cwd=str(root / "updater"),
+            capture_output=True,
+            text=True,
+            timeout=120,
+        )
+        if build.returncode != 0:
+            pytest.skip(f"go build lockhelper failed: {build.stderr}")
+
+        lock_path = tmp_path / "config" / "locks" / "runtime.lock"
+        lock_path.parent.mkdir(parents=True)
+
+        # Direction 1: Python holds, Go tries
+        plock = AdvisoryFileLock(lock_path)
+        plock.acquire()
+        try:
+            proc = subprocess.run(
+                [str(helper), "try", "-path", str(lock_path)],
+                capture_output=True,
+                text=True,
+                timeout=15,
+            )
+            assert proc.returncode == 2
+            assert "busy" in proc.stdout.lower()
+        finally:
+            plock.release()
+
+        # Direction 2: Go holds, Python tries
+        hold = subprocess.Popen(
+            [str(helper), "hold", "-path", str(lock_path), "-seconds", "5"],
+            stdout=subprocess.PIPE,
+            stderr=subprocess.PIPE,
+            text=True,
+        )
+        try:
+            # wait until held
+            import time
+
+            deadline = time.time() + 5
+            held = False
+            while time.time() < deadline:
+                line = hold.stdout.readline() if hold.stdout else ""
+                if "held" in line.lower():
+                    held = True
+                    break
+            assert held, "lockhelper did not report held"
+            from services.process_lock import LockBusyError
+
+            with pytest.raises(LockBusyError):
+                AdvisoryFileLock(lock_path).acquire(timeout_seconds=None)
+        finally:
+            hold.terminate()
+            try:
+                hold.wait(timeout=5)
+            except Exception:
+                hold.kill()
