+++ tests/unit/test_system_scheduler_backend.py
@@ -1,5 +1,6 @@
 """系统级调度后端单元测试（Phase 2 capability + golden artifacts）。"""
 
+import xml.etree.ElementTree as ET
 from datetime import datetime, timedelta, timezone
 from pathlib import Path
 
@@ -28,6 +29,8 @@ from services.system_scheduler_backend import (
     windows_join_args,
     windows_quote_argument,
 )
+
+TID = "550e8400-e29b-41d4-a716-446655440000"
 
 
 class TestValidateTaskId:
@@ -320,3 +323,325 @@ class TestGoldenArtifacts:
         )
         # command body uses shlex quotes
         assert "&&" in line
+
+
+# ---------------------------------------------------------------------------
+# Capabilities smoke evidence
+# ---------------------------------------------------------------------------
+
+
+class TestCapabilitiesSmoke:
+    def test_no_evidence_all_disabled(self):
+        caps = build_capabilities("windows", host_platform="windows")
+        assert all(not c.enabled for c in caps.cells)
+
+    def test_explicit_smoke_enables_user_trigger_specific(self):
+        caps = build_capabilities(
+            "windows",
+            host_platform="windows",
+            smoke_evidence={
+                "windows:user:cron": True,
+                "windows:user:date": True,
+                "windows:user:interval": True,
+            },
+        )
+        user = [
+            c
+            for c in caps.cells
+            if c.platform == "windows" and c.scope == SystemTaskScope.USER
+        ]
+        assert all(c.enabled for c in user)
+        system = [
+            c
+            for c in caps.cells
+            if c.platform == "windows" and c.scope == SystemTaskScope.SYSTEM
+        ]
+        assert all(not c.enabled for c in system)
+
+    def test_broad_windows_evidence_does_not_enable(self):
+        caps = build_capabilities(
+            "windows",
+            host_platform="windows",
+            smoke_evidence={"windows": True, "windows:user": True},
+        )
+        assert all(not c.enabled for c in caps.cells)
+
+    def test_string_false_not_truthy(self, tmp_path: Path):
+        cfg = tmp_path / "config"
+        cfg.mkdir()
+        (cfg / "system_scheduler_smoke.json").write_text(
+            '{"windows:user:cron": "false", "windows:user:date": true}',
+            encoding="utf-8",
+        )
+        from services.system_scheduler_backend import load_smoke_evidence
+
+        ev = load_smoke_evidence(tmp_path)
+        assert "windows:user:cron" not in ev  # string rejected
+        assert ev.get("windows:user:date") is True
+
+    def test_system_requires_elevated_proof(self):
+        caps = build_capabilities(
+            "windows",
+            host_platform="windows",
+            smoke_evidence={
+                "windows:system:cron": True,  # insufficient alone
+            },
+        )
+        sys_cron = next(
+            c
+            for c in caps.cells
+            if c.platform == "windows"
+            and c.scope == SystemTaskScope.SYSTEM
+            and c.trigger_type == "cron"
+        )
+        assert sys_cron.enabled is False
+        caps2 = build_capabilities(
+            "windows",
+            host_platform="windows",
+            smoke_evidence={
+                "windows:system:cron": True,
+                "windows:system:elevated": True,
+                "windows:system:post_user_restart": True,
+            },
+        )
+        sys_cron2 = next(
+            c
+            for c in caps2.cells
+            if c.platform == "windows"
+            and c.scope == SystemTaskScope.SYSTEM
+            and c.trigger_type == "cron"
+        )
+        assert sys_cron2.enabled is True
+
+
+# ---------------------------------------------------------------------------
+# Windows XML verification
+# ---------------------------------------------------------------------------
+
+
+class TestWindowsXmlVerify:
+    def _spec(self, scope=SystemTaskScope.USER, trig="cron") -> SystemTaskSpec:
+        if trig == "cron":
+            trigger = OSTriggerSpec(trigger_type="cron", cron_expression="0 9 * * *")
+        elif trig == "interval":
+            trigger = OSTriggerSpec(trigger_type="interval", interval_minutes=15)
+        else:
+            trigger = OSTriggerSpec(
+                trigger_type="date",
+                run_date=datetime.now(timezone.utc) + timedelta(days=2),
+            )
+        return SystemTaskSpec(
+            task_id=TID,
+            task_name="T",
+            exe_path=r"C:\Program Files\MWU\python.exe",
+            cli_args=[r"C:\Program Files\MWU\main.py", "--headless", "--task", TID],
+            trigger=trigger,
+            scope=scope,
+            working_dir=r"C:\Program Files\MWU",
+        )
+
+    def test_golden_xml_verifies(self):
+        b = WindowsBackend()
+        spec = self._spec()
+        raw = b._build_task_xml(spec)
+        ok, detail = b.compare_exported_xml_bytes(raw, spec)
+        assert ok, detail
+
+    def test_wrong_command_fails(self):
+        b = WindowsBackend()
+        spec = self._spec()
+        raw = (
+            b._build_task_xml(spec)
+            .decode("utf-8")
+            .replace(spec.exe_path, r"C:\wrong\python.exe")
+        )
+        ok, detail = b.compare_exported_xml_bytes(raw.encode("utf-8"), spec)
+        assert not ok
+        assert "Command" in detail
+
+    def test_wrong_args_fails(self):
+        b = WindowsBackend()
+        spec = self._spec()
+        bad = self._spec()
+        bad.cli_args = ["other.py"]
+        raw = b._build_task_xml(bad)
+        ok, detail = b.compare_exported_xml_bytes(raw, spec)
+        assert not ok
+        assert "Arguments" in detail
+
+    def test_serviceaccount_fails(self):
+        b = WindowsBackend()
+        spec = self._spec(scope=SystemTaskScope.SYSTEM)
+        raw = (
+            b._build_task_xml(spec)
+            .decode("utf-8")
+            .replace("Password", "ServiceAccount")
+        )
+        ok, detail = b.compare_exported_xml_bytes(raw.encode("utf-8"), spec)
+        assert not ok
+
+    def test_pt1m_cron_fails(self):
+        b = WindowsBackend()
+        spec = self._spec(trig="cron")
+        # inject PT1M into golden
+        root = ET.fromstring(b._build_task_xml(spec))
+        # append bogus repetition
+        for el in root.iter():
+            if el.tag.endswith("CalendarTrigger") or el.tag == "CalendarTrigger":
+                rep = ET.SubElement(el, "Repetition")
+                iv = ET.SubElement(rep, "Interval")
+                iv.text = "PT1M"
+                break
+        raw = ET.tostring(root, encoding="utf-8")
+        ok, detail = b.compare_exported_xml_bytes(raw, spec)
+        assert not ok
+        assert "PT1M" in detail
+
+    def test_cron_0900_to_1000_fails(self):
+        b = WindowsBackend()
+        spec = self._spec(trig="cron")  # 0 9 * * *
+        root = ET.fromstring(b._build_task_xml(spec))
+        for el in root.iter():
+            if el.tag.endswith("StartBoundary") or el.tag == "StartBoundary":
+                if el.text:
+                    el.text = el.text[:11] + "10" + el.text[13:]
+                break
+        ok, detail = b.compare_exported_xml_bytes(
+            ET.tostring(root, encoding="utf-8"), spec
+        )
+        assert not ok
+        assert "StartBoundary" in detail or "time mismatch" in detail
+
+    def test_date_drift_fails(self):
+        b = WindowsBackend()
+        spec = self._spec(trig="date")
+        root = ET.fromstring(b._build_task_xml(spec))
+        for el in root.iter():
+            if el.tag.endswith("StartBoundary") or el.tag == "StartBoundary":
+                if el.text and len(el.text) >= 10:
+                    # shift day
+                    el.text = (
+                        el.text[:8] + ("2" if el.text[8] != "2" else "3") + el.text[9:]
+                    )
+                break
+        ok, detail = b.compare_exported_xml_bytes(
+            ET.tostring(root, encoding="utf-8"), spec
+        )
+        assert not ok
+        assert "StartBoundary" in detail
+
+    def test_days_interval_change_fails(self):
+        b = WindowsBackend()
+        spec = self._spec(trig="cron")
+        root = ET.fromstring(b._build_task_xml(spec))
+        for el in root.iter():
+            if el.tag.endswith("DaysInterval") or el.tag == "DaysInterval":
+                el.text = "2"
+                break
+        ok, detail = b.compare_exported_xml_bytes(
+            ET.tostring(root, encoding="utf-8"), spec
+        )
+        assert not ok
+        assert "DaysInterval" in detail
+
+    def test_settings_change_fails(self):
+        b = WindowsBackend()
+        spec = self._spec()
+        root = ET.fromstring(b._build_task_xml(spec))
+        for el in root.iter():
+            if el.tag.endswith("ExecutionTimeLimit") or el.tag == "ExecutionTimeLimit":
+                el.text = "PT1H"
+                break
+        ok, detail = b.compare_exported_xml_bytes(
+            ET.tostring(root, encoding="utf-8"), spec
+        )
+        assert not ok
+        assert "ExecutionTimeLimit" in detail
+
+    def test_settings_enabled_false_with_trigger_enabled_true_fails(self):
+        """P0-1: _find_desc hit Trigger/Enabled first, bypassing Settings/Enabled=false."""
+        b = WindowsBackend()
+        spec = self._spec(trig="cron")
+        root = ET.fromstring(b._build_task_xml(spec))
+        # Flip Settings/Enabled to false; verify must fail (subtree lookup)
+        for el in root:
+            if b._local_tag(el.tag) == "Settings":
+                for child in el:
+                    if b._local_tag(child.tag) == "Enabled":
+                        child.text = "false"
+                break
+        ok, detail = b.compare_exported_xml_bytes(
+            ET.tostring(root, encoding="utf-8"), spec
+        )
+        assert not ok
+        assert "Settings.Enabled" in detail or "true" in detail.lower()
+
+    def test_allow_start_on_demand_verified(self):
+        """P0-2: AllowStartOnDemand is built but must be verified."""
+        b = WindowsBackend()
+        spec = self._spec()
+        root = ET.fromstring(b._build_task_xml(spec))
+        # Golden passes
+        ok, _ = b.compare_exported_xml_bytes(ET.tostring(root, encoding="utf-8"), spec)
+        assert ok
+        # Mangle AllowStartOnDemand -> fails
+        for el in root.iter():
+            if b._local_tag(el.tag) == "AllowStartOnDemand":
+                el.text = "false"
+                break
+        ok2, detail2 = b.compare_exported_xml_bytes(
+            ET.tostring(root, encoding="utf-8"), spec
+        )
+        assert not ok2
+        assert "AllowStartOnDemand" in detail2
+
+
+# ---------------------------------------------------------------------------
+# Linux bounds / macOS escaping
+# ---------------------------------------------------------------------------
+
+
+class TestLinuxMacStrict:
+    def test_linux_minute_out_of_bounds(self):
+        with pytest.raises(ValueError, match="越界"):
+            validate_linux_cron_expression("60 9 * * *")
+
+    def test_linux_hour_out_of_bounds(self):
+        with pytest.raises(ValueError, match="越界"):
+            validate_linux_cron_expression("0 24 * * *")
+
+    def test_linux_day_out_of_bounds(self):
+        with pytest.raises(ValueError, match="越界"):
+            validate_linux_cron_expression("0 9 32 * *")
+
+    def test_linux_invalid_expansions_rejected(self):
+        with pytest.raises(ValueError):
+            validate_linux_cron_expression("70/2 9 * * *")
+        with pytest.raises(ValueError):
+            validate_linux_cron_expression("0 9 32/2 * *")
+        with pytest.raises(ValueError):
+            validate_linux_cron_expression("5-3 9 * * *")
+        with pytest.raises(ValueError):
+            validate_linux_cron_expression("*/0 9 * * *")
+
+    def test_macos_admin_script_uses_base64(self):
+        b = MacOSBackend()
+        script = b._admin_register_script(
+            "/Library/LaunchDaemons/com.mwu.daemon.x.plist",
+            "com.mwu.daemon.x",
+            '<?xml version="1.0"?><plist><dict><key>Label</key><string>a\'b"c</string></dict></plist>',
+        )
+        assert "base64" in script
+        assert "echo '" not in script  # no raw echo of plist
+        assert "a'b" not in script
+
+    def test_macos_osascript_uses_argv_not_double_quoted_literal(self):
+        b = MacOSBackend()
+        # Inspect the generated elevation invocation pattern via source of method
+        import inspect
+
+        src = inspect.getsource(b._run_osascript_admin)
+        assert "on run argv" in src
+        assert "item 1 of argv" in src
+        # must not build do shell script "..." with embedded script
+        assert 'do shell script "' not in src or "item 1" in src
